requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
python-socketio==5.8.0
python-engineio==4.7.1
Werkzeug==2.3.7
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
redis==4.6.0
celery==5.3.2
gunicorn==21.2.0
requests==2.31.0
Pillow==10.0.0
boto3==1.28.57
python-dotenv==1.0.0
marshmallow==3.20.1
webargs==8.3.0
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# Machine Learning dependencies
scikit-learn==1.3.0
numpy==1.25.2
pandas==2.1.0
textblob==0.17.1
nltk==3.8.1

# Image processing
opencv-python==4.8.0.76
pytesseract==0.3.10

# Development tools
python-decouple==3.8
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1
